semi-working usb mass storage
updated the kernel with the .32 usb_composite drivers, so now the
adb notification comes up as well as usb mass storage works.  the
only bug is that this will not occur if you are plugged in while the
system turns on.  you must plug in AFTER the system has come up for
it to work, need to figure out why.

Change-Id: I2eeb834f4ef7778d3356dbd2690b06048df3661b