Fix reseting the bcb for stock 1.1.0 uboot.bin
The old way of writing four block of zeros to /rom/bcb doesn't work with the new
u-boot.bin included with the stock 1.1.0 update which results in the device
always reboot into recovery.

This writes 1088 zeros to /rom/bcb which was copied from the romrestore.zip also
included with the 1.1.0 stock update.

Change-Id: Ic174f273a4dc27a0351212a27ca6b697a30a3649