Change /system from ext2 to ext3
Even though the stock rom mounts /system as ext2, formatting it as ext3 will be
better.   Mounting an ext3 fs as ext2 will just ignore the journaling which
will be ok for us.  Also recovery should follow the format in the edify
scripting, so this is only for corner cases anyway.

Change-Id: I3beebc85360d714f5ce43898957bea8deaa3580f