Add primary and secondary bootloaders to the update package
This is the initial addition of MLO and u-boot to the update package.

Since the new kernel has dependancies on proper versions of MLO and u-boot
we are going to bundle them with the update.zip which should help out
all users of CM for encore.

Users who are used to a custom uboot.bin will have to reflash their u-boot.bin
after flashing a CM7 update.  This is analogous to custom OC kernels.

This is the initial u-boot.bin which doesn't have many features.  Future
version of the CM bundled u-boot binary will have additional features
which will probably make the use of custom u-boot binaries unnecessary.

This requires http://review.cyanogenmod.com/#change,5489

Change-Id: I51216e1afc6477bac2d7aaa04a678a411df460fd