Set ro.serialno at boot
This follows the hack found in geeksphone one and the commtiva z71
to set the ro.serialno shortly after init has ran.

Currently ro.serialno is not being set from /proc/cmdline and is
returning an emptry string

This is to support an effort by Swype to support devies that return
a null getDeviceId.  Hopefully other programs will follow suit.

Change-Id: I04c23ec9e3fd6cf400631d18d187b05938a8b44b