Make /dev/ion 0660
This file appears to be 0777 on tuna, but it looks like it's only used
by the graphics subsystem (surfaceflinger, running as user/group system)
and the media subsystem (mediaserver, running as user media).  With this
in mind, set permissions on /dev/ion to 0660, the owner as media, and
group as system.