Symlink /system/app/{GoogleHome,Hangouts}.apk to /dev/null
GoogleHome is included in some Google apps packages for some reason, but
doesn't fit onto our small /system partition.  Ship a symlink to
/dev/null there, which will cause the OTA installer to write the package
to the bit bucket.  (It's still possible to install the package via the
usual methods -- we just don't want it on /system.)

Do the same with Hangouts, which can be installed from the Market.

Thanks to Ricardo Cerqueira (arcee) for the idea.