Report only hardware capabilities that we actually have
We're currently reporting a bunch of hardware capabilities (camera,
microphone, compass, GPS) which we don't actually have.  Presumably this
was done because these capabilities were required by the compatibility
definition once upon a time and apps implicitly depended on their
presence, but with the revision of the CDD to accommodate basic
wifi-only tablets and the proliferation of such devices on the market,
this should no longer be necessary.