Enable KSM by default
Kernel Samepage Merging (KSM) periodically scans memory marked by the
framework as MADV_MERGEABLE looking for duplicate pages; when it finds
them, it drops the duplicates and points all users at a single copy of
the page.  This is particularly useful on Android, where much of the
Java class data can be shared across processes.  Enabling this feature
on encore can save 16 MB of RAM after just a few hours of uptime.

Change-Id: Ic4eed0e0e42b3170e55831e4c85373b8720ba29a