Recovery: display whether operating on eMMC or SD on startup
Particularly with the new location-agnostic recoveries, it's possible to
get confused as to whether backups/restores, wipes, and (eventually) OTA
flashes will affect the eMMC or SD card.  Detect whether the recovery is
operating on eMMC or SD on startup and display this to the user to try
to avoid mixups.